requirements.txt
# ─── Core utilities ─────────────────────────────────────────
requests>=2.31
python-dotenv>=1.0
tqdm>=4.66
rich>=13.7               # nice CLI progress / tables

# ─── Data wrangling & science ──────────────────────────────
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4

# ─── Deep-learning stack (CPU or CUDA) ─────────────────────
torch>=2.2               # PyTorch
torchmetrics>=1.3
einops>=0.7              # lightweight tensor ops helper

# ─── Orbital mechanics ─────────────────────────────────────
sgp4>=2.23

# ─── Visualisation & notebooks ─────────────────────────────
matplotlib>=3.9
seaborn>=0.13
jupyterlab>=4.2
ipykernel>=6.29

# ─── Optional, install on demand ───────────────────────────
# transformers>=4.40       # for TimeGPT or HF models
# nni, pytorch-wavelets, yarl … (Informer repos sometimes need these)
